% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{bounded.isoreg}
\alias{bounded.isoreg}
\title{Bounded isotonic regression}
\usage{
bounded.isoreg(y.isoreg, y.mean)
}
\arguments{
\item{y.isoreg}{A numeric vector of the isotonic regressor.}

\item{y.mean}{A number of the sample mean of the observations.}
}
\value{
A list with components:
\describe{
\item{a}{A numeric vector of the lower range estimates along
the regularization path.}
\item{b}{A numeric vector of the upper range estimates along
the regularization path.}
\item{lambda}{A numeric vector of the regularization parameters
along the regularization path.}
}
}
\description{
Bounded isotonic regression
}
\references{
Luss, R. and Rosset, S. (2017)
\emph{Bounded isotonic regression, Electronic Journal of Statistics, Vol. 11, 4488-4514},
\doi{10.1214/17-EJS1365}.
}
