% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{em.estep}
\alias{em.estep}
\title{Compute the conditional expectation of uncured status in the E-step of
the EM-algorithm}
\usage{
em.estep(y, delta, p, s.u, y.max = max(y[delta == 1]))
}
\arguments{
\item{y}{A numeric vector (length \code{n}) of the observed event time.}

\item{delta}{A numeric vector (length \code{n}) of the censoring indicator.}

\item{p}{A numeric vector (length \code{n}) of the uncured probability values.}

\item{s.u}{A numeric vector (length \code{n}) of the
conditional survival function values, see \code{\link{estimate.survival}}.}

\item{y.max}{A positive number of the cure threshold.
Default is the maximum uncensored event time.}
}
\value{
A numeric vector of the conditional expected uncured status.
}
\description{
Compute the conditional expectation of uncured status in the E-step of
the EM-algorithm
}
\keyword{internal}
