% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{em.link.bounded.isoreg}
\alias{em.link.bounded.isoreg}
\title{EM algorithm for monotone link estimator using bounded isotonic regression}
\usage{
em.link.bounded.isoreg(
  y,
  delta,
  index,
  uncured.survival,
  rev.km.fit,
  p.init,
  y.max = max(y[delta == 1]),
  n.fold = 10,
  truncation.lower = 1e-06,
  truncation.upper = 1 - 1e-06,
  tol = 1e-06,
  max.iter = 1L,
  shuffle = FALSE
)
}
\arguments{
\item{y}{A numeric vector (length \code{n}) of the observed event time.}

\item{delta}{A numeric vector (length \code{n}) of the censoring indicator.}

\item{index}{A numeric vector (length \code{n}) of index, \eqn{\gamma^TX}.}

\item{uncured.survival}{A numeric vector (length \code{n}) of the
conditional survival function values, see \code{\link{estimate.survival}}.}

\item{rev.km.fit}{A \code{\link[stats]{stepfun}} of the reverse Kaplan-Meier estimate from
\code{\link{estimate.km}}.}

\item{p.init}{A numeric vector (length \code{n}) of the
initial uncured probability values.}

\item{y.max}{A positive number of the cure threshold.
Default is the maximum uncensored event time.}

\item{n.fold}{An integer of the number folds used in the cross-validation for
determining the truncation. Default is \code{10}.}

\item{truncation.lower}{A positive number of the lower truncation parameter for the link.
Default is \code{1e-6}.}

\item{truncation.upper}{A positive number of the upper truncation parameter for the link.
Default is \code{1 - 1e-6}.}

\item{tol}{A positive number of convergence tolerance. Default is \code{1e-6}}

\item{max.iter}{A positive integer the of maximum number of iterations of the EM algorithm.
Default is \code{1L}.}

\item{shuffle}{A logical if data is shuffled. Default is \code{FALSE}.}
}
\value{
A list with components:
\describe{
\item{link}{A \code{\link[stats]{stepfun}} of the link estimate.}
\item{convergence}{A logical if the EM algorithm converged.}
}
}
\description{
EM algorithm for monotone link estimator,
    \eqn{\hat{\varphi}_{n, \theta}}, using bounded isotonic regression to
    determine the truncation.
}
