% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{em.link}
\alias{em.link}
\title{EM algorithm for monotone link estimator}
\usage{
em.link(
  y,
  delta,
  index,
  uncured.survival,
  p.init,
  y.max = max(y[delta == 1]),
  truncation.lower = 1e-06,
  truncation.upper = 1 - 1e-06,
  tol = 1e-06,
  max.iter = 50L
)
}
\arguments{
\item{y}{A numeric vector (length \code{n}) of the observed event time.}

\item{delta}{A numeric vector (length \code{n}) of the censoring indicator.}

\item{index}{A numeric vector (length \code{n}) of index, \eqn{\gamma^TX}.}

\item{uncured.survival}{A numeric vector (length \code{n}) of the
conditional survival function values, see \code{\link{estimate.survival}}.}

\item{p.init}{A numeric vector (length \code{n}) of the
initial uncured probability values.}

\item{y.max}{A positive number of the cure threshold.
Default is the maximum uncensored event time.}

\item{truncation.lower}{A positive number of the lower truncation parameter
for the link. Default is \code{1e-6}.}

\item{truncation.upper}{A positive number of the upper truncation parameter
for the link. Default is \code{1 - 1e-6}.}

\item{tol}{A positive number of convergence tolerance.}

\item{max.iter}{A positive integer of maximum number of iterations.}
}
\value{
A list with components:
\describe{
\item{link}{A \code{\link[stats]{stepfun}} of the monotone link estimate.}
\item{convergence}{A logical if the EM algorithm converges.}
}
}
\description{
EM algorithm for monotone link estimator, \eqn{\hat{\varphi}_{n, \theta}}.
}
\keyword{internal}
