% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{epecp}
\alias{epecp}
\title{Expected prediction error for cure probability (EPECP)}
\usage{
epecp(y, delta, p.est, km.stepfun, y.max = max(y[delta == 1]), n = length(y))
}
\arguments{
\item{y}{A numeric vector (length \code{n}) of the observed event time.}

\item{delta}{A numeric vector (length \code{n}) of the censoring indicator.}

\item{p.est}{A numeric vector (length \code{n}) of the
estimated uncured probability.}

\item{y.max}{A positive number of the cure threshold.
Default is the maximum uncensored event time.}

\item{n}{An integer of the sample size.}

\item{revkm.stepfun}{A \code{\link[stats]{stepfun}} of the reversed Kaplan-Meier estimate of
the survivor function, see \code{\link{estimate.km}}.}
}
\value{
A numeric of the computed EPECP.
}
\description{
Expected prediction error for cure probability (EPECP)
}
\references{
Jiang, W., Sun, H. and Peng, Y. (2017)
\emph{Prediction accuracy for the cure probabilities in mixture cure models,
Statistical Methods in Medical Research, Vol. 26, 2029-2041},
\doi{10.1177/0962280217708673}.
}
\keyword{internal}
