% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latency.R
\name{estimate.beta}
\alias{estimate.beta}
\title{Estimate the latency regression parameter}
\usage{
estimate.beta(y, delta, Z, w, beta.init = rep(0, ncol(Z)))
}
\arguments{
\item{y}{A numeric vector (length \code{n}) of the observed event time.}

\item{delta}{A numeric vector (length \code{n}) of the censoring indicator.}

\item{Z}{A numeric matrix (dimension \code{n} x \code{q};
each row is an observation vector) of covariates for the latency (no intercept term).}

\item{w}{A numeric vector (length \code{n}) of the
conditional expected uncured status from the E-step, see \code{\link{em.estep}}.}

\item{beta.init}{A numeric vector (length \code{q}) of the
initial latency regression parameter value. Default is a zero vector.}
}
\value{
A list with components: \describe{
  \item{beta}{A vector (length \code{q}) of the latency regression parameter estimate.}
  \item{fit}{An object returned from \code{\link[survival]{coxph}}.}
}
}
\description{
Estimate the latency regression parameter, \eqn{\beta}.
}
\keyword{internal}
