% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{estimate.incidence}
\alias{estimate.incidence}
\title{Maximize the log-likelihood of the incidence part w.r.t. regression parameter
using the Augmented Lagrangian method}
\usage{
estimate.incidence(
  X,
  y,
  delta,
  w,
  s.u.current,
  p.init,
  gamma.init,
  bw,
  y.max,
  truncation.lower,
  truncation.upper
)
}
\arguments{
\item{X}{A numeric matrix (dimension \code{n} x \code{d};
each row is an observation vector) of covariates for the incidence
(no intercept term).}

\item{y}{A numeric vector (length \code{n}) of the observed event time.}

\item{delta}{A numeric vector (length \code{n}) of the censoring indicator.}

\item{w}{A numeric vector (length \code{n}) of the
conditional expected uncured status from the E-step, see \code{\link{em.estep}}.}

\item{s.u.current}{A numeric vector (length \code{n}) of the
conditional survival function values, see \code{\link{estimate.survival}}.}

\item{p.init}{A numeric vector (length \code{n}) of the
initial uncured probability values.}

\item{gamma.init}{A numeric vector (length \code{d}) of the
initial incidence regression parameter value.}

\item{bw}{A positive number of the bandwidth parameter for smoothing.}

\item{y.max}{A positive number of the cure threshold.}

\item{truncation.lower}{A positive number of the lower truncation parameter for the link.}

\item{truncation.upper}{A positive number of the upper truncation parameter for the link.}
}
\value{
A list with components:
\describe{
\item{gamma}{A numeric vector (length \code{d}) of the incidence
regression parameter estimate.}
\item{convergence}{A integer code returned by \code{\link[Rsolnp]{solnp}}.}
\item{optim}{A list returned by \code{\link[Rsolnp]{solnp}}}
\item{l}{A \code{\link[stats]{stepfun}} of the monotone link estimate from the EM-algorithm,
see \code{\link{em.link}}.}
\item{smoothed.link}{A function of the smoothed monotone link estimate,
see \code{\link{smooth.monotone.link}}.}
}
}
\description{
Maximize the log-likelihood of the incidence part w.r.t. \eqn{\gamma}.
}
\keyword{internal}
