% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate.km}
\alias{estimate.km}
\title{Kaplan-Meier estimate of survivor function}
\usage{
estimate.km(y, delta)
}
\arguments{
\item{y}{A numeric vector (length \code{n}) of the observed event time.}

\item{delta}{A numeric vector (length \code{n}) of the censoring indicator.}
}
\value{
A \code{\link[stats]{stepfun}} of the estimate.
}
\description{
Kaplan-Meier estimate of survivor function
}
