% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latency.R
\name{estimate.survival}
\alias{estimate.survival}
\title{Estimate the conditional survival function for the latency}
\usage{
estimate.survival(y, delta, Z, beta, w)
}
\arguments{
\item{delta}{A numeric vector (length \code{n}) of the censoring indicator.}

\item{Z}{A numeric matrix (dimension \code{n} x \code{q};
each row is an observation vector) of covariates for the latency (no intercept term).}

\item{beta}{A numeric vector (length \code{q}) of the
initial latency regression parameter value.}

\item{w}{A numeric vector (length \code{n}) of the
conditional expected uncured status from the E-step, see \code{\link{em.estep}}.}

\item{A}{numeric vector (length \code{n}) of the observed event time.}
}
\value{
A list with components:
\describe{
    \item{s.u}{A vector of the conditional survival function values.}
    \item{s.baseline}{A vector of the conditional baseline survival function values.}
    \item{baseline}{The list return from \code{\link{baseline.hazard}}}
}
}
\description{
Estimate the conditional survival function for the latency, \eqn{S_u(t|Z)}.
}
\keyword{internal}
