% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{integrated.kern}
\alias{integrated.kern}
\title{Integrated tri-weight kernel function}
\usage{
integrated.kern(x)
}
\arguments{
\item{x}{A numeric vector of points for which the value of
`integrated.kern` is computed.}
}
\value{
A numeric vecotr of the computed `integrated.kern`.
}
\description{
Integrated tri-weight kernel function:
\deqn{\int_{-\infty}^x k(u)du,}
where \eqn{k(u) = \frac{35}{32}(1-u^2)^3I_{[-1,1]}(u)}.
}
\keyword{internal}
