% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{isotonized.link}
\alias{isotonized.link}
\title{Fit a isotonic regression for the monotone link estimator}
\usage{
isotonized.link(n, index, w)
}
\arguments{
\item{n}{An integer of the sample size.}

\item{index}{A numeric vector (length \code{n}) of index computed using all
observations for a fixed incidence regression parameter, \eqn{\gamma}.}

\item{w}{A numeric vector (length \code{n}) of the
conditional expected uncured status from the E-step, see \code{\link{em.estep}}.}
}
\value{
An object of class \code{\link[stats]{isoreg}}.
}
\description{
Fit a isotonic regression for the monotone link estimator
}
\keyword{internal}
