% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kfold}
\alias{kfold}
\title{Split data into K folds}
\usage{
kfold(n, k, shuffle = FALSE)
}
\arguments{
\item{n}{An integer of sample size.}

\item{k}{An integer of number of folds.}

\item{shuffle}{A logical if data is shuffled.}
}
\value{
A list of indices of each fold.
}
\description{
Split data into K folds
}
\keyword{internal}
