% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence.R
\name{llh.incidence}
\alias{llh.incidence}
\title{Log-likelihood function of the incidence part}
\usage{
llh.incidence(p, w)
}
\arguments{
\item{p}{A numeric vector (length \code{n}) of the uncured probabilities.}

\item{w}{A numeric vector (length \code{n}) of the
conditional expected uncured status from the E-step, see \code{\link{em.estep}}.}
}
\value{
The computed log-likelihood value.
}
\description{
Log-likelihood function of the incidence part:
    \deqn{\sum_{i=1}^{n}w_i\log(p_i) + (1-w_i)\log(1-p_i).}
}
\keyword{internal}
