% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msic.R
\name{msic}
\alias{msic}
\title{Fit a monotone single-index mixture cure model}
\usage{
msic(
  X,
  Z,
  y,
  delta,
  bw = NULL,
  truncation.lower = 1e-06,
  truncation.upper = 1 - 1e-06,
  max.iter = 50L,
  tol = 1e-05,
  truncation.est = FALSE,
  gamma.init = NULL,
  beta.init = NULL
)
}
\arguments{
\item{X}{A numeric matrix (dimension \code{n} x \code{d};
each row is an observation vector) of covariates for the incidence (no intercept term).}

\item{Z}{A numeric matrix (dimension \code{n} x \code{q};
each row is an observation vector) of covariates for the latency (no intercept term).}

\item{y}{A numeric vector (length \code{n}) of the observed event time.}

\item{delta}{A numeric vector (length \code{n}) of the censoring indicator.}

\item{bw}{A positive number of bandwidth, \code{NULL} indicates using
\eqn{rn^{-1/5}}, where \eqn{r} is the range of the index \eqn{\gamma^TX}.}

\item{truncation.lower}{A positive number of the lower truncation parameter for the link.
Default is \code{1e-6}. See Details below.}

\item{truncation.upper}{A positive number of the upper truncation parameter for the link.
Default is \code{1 - 1e-6}. See Details below.}

\item{max.iter}{An integer of the maximum number of iterations.
Default is \code{50L}.}

\item{tol}{A positive number of convergence tolerance. Default is \code{1e-5}.}

\item{truncation.est}{A logical if the upper and lower truncation are to be
determined by the bounded isotonic regression,
see \code{\link{em.link.bounded.isoreg}}. See also Details below.}

\item{gamma.init}{A numeric vector (length \code{d}) of the
initial incidence regression parameter value.
\code{NULL} indicates it is initialized by fitting a logistic regression of
the censoring indicator (\code{delta}) against the covariates \code{X}.
Default is \code{NULL}.}

\item{beta.init}{A numeric vector (length \code{q}) of the
initial latency regression parameter value.
\code{NULL} indicates it is initialized by fitting a Cox PH model
\code{\link[survival]{coxph}} to the uncensored observations with covariates \code{Z}.
Default is \code{NULL}.}
}
\value{
An object of class \code{msic}.
}
\description{
Fit a monotone single-index mixture cure model
}
\details{
If \code{truncation.lower} or \code{truncation.upper} is \code{NULL},
then the two truncation parameters are determined from
the logistic regression of the censoring indicator, \code{delta}, against
the covariates, \code{X}, in the initialization step.
If \code{truncation.est} is \code{TRUE}, the two truncation parameters are
determined by the bounded isotonic regression, see \code{\link{em.link.bounded.isoreg}}.
}
\examples{
# Using default truncation and simulated Data 1 ----------------------------
data("data1")
X <- cbind(data1$X1, data1$X2, data1$X3, data1$X4)
Z <- cbind(data1$Z1, data1$Z2)
# Model estimation
m.fit.1 <- msic(X, Z, data1$y, data1$delta, truncation.est = FALSE)
# Print the estimated regression parameters
print(m.fit.1)
# Plot the estimated link functions
plot(m.fit.1)
# Predict uncured probability
X.new <- cbind(
  matrix(runif(5), ncol = 1),
  matrix(rnorm(5, 0, 1), ncol = 1),
  matrix(rbinom(5, 1, 0.3), ncol = 1),
  matrix(rbinom(5, 1, 0.6), ncol = 1)
)
p.hat.1 <- predict(m.fit.1, X.new)

# Using truncation being determined by the bounded isotonic regression
# method and simulated Data 2 ----------------------------
data("data2")
X <- cbind(data2$X1, data2$X2, data2$X3, data2$X4)
Z <- cbind(data2$Z1, data2$Z2)
# Model estimation
m.fit.2 <- msic(X, Z, data2$y, data2$delta, truncation.est = TRUE)
# Print the estimated regression parameters
print(m.fit.2)
# Plot the estimated link functions
plot(m.fit.2)
# Predict uncured probability
X.new <- cbind(
  matrix(runif(5), ncol = 1),
  matrix(rnorm(5, 0, 1), ncol = 1),
  matrix(rbinom(5, 1, 0.3), ncol = 1),
  matrix(rbinom(5, 1, 0.6), ncol = 1)
)
p.hat.2 <- predict(m.fit.2, X.new)

# Using pre-specified truncation and simulated Data 2 ----------------------------
data("data2")
X <- cbind(data2$X1, data2$X2, data2$X3, data2$X4)
Z <- cbind(data2$Z1, data2$Z2)
# Model estimation with pre-specified truncation
m.fit.3 <- msic(X, Z, data2$y, data2$delta, truncation.est = FALSE,
                truncation.lower = 0.5, truncation.upper = 0.75)
# Print the estimated regression parameters
print(m.fit.3)
# Plot the estimated link functions
plot(m.fit.3)
# Predict uncured probability
X.new <- cbind(
  matrix(runif(5), ncol = 1),
  matrix(rnorm(5, 0, 1), ncol = 1),
  matrix(rbinom(5, 1, 0.3), ncol = 1),
  matrix(rbinom(5, 1, 0.6), ncol = 1)
)
p.hat.3 <- predict(m.fit.3, X.new)

}
\references{
Musta, E. and Yuen, T. P. (2022)
\emph{Single-index mixture cure model under monotonicity constraints, arXiv preprint},
\doi{10.48550/ARXIV.2211.09464}.
}
