% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize.gamma}
\alias{normalize.gamma}
\title{Normalizing incidence regression parameter}
\usage{
normalize.gamma(g)
}
\arguments{
\item{g}{A numeric vector (length \code{d}) of the incidence regression parameter.}
}
\value{
A numeric vector (length \code{d}) of the normalized regression parameter.
}
\description{
Normalizing incidence regression parameter
}
\keyword{internal}
