% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msic.R
\name{plot.msic}
\alias{plot.msic}
\title{Plot a msic object}
\usage{
\method{plot}{msic}(x, ...)
}
\arguments{
\item{x}{An \code{msic} object.}

\item{\dots}{Additional plot arguments.}
}
\value{
A plot mentioned in Details.
}
\description{
Plot the estimated monotone link functions.
}
\details{
The estimated monotone link function \eqn{\hat{\varphi}_{n,\theta}} and
the smoothed monotone link function \eqn{\hat{\varphi}_{n,\theta}^s} are plotted.
}
