% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msic.R
\name{predict.msic}
\alias{predict.msic}
\title{Predict uncure probability}
\usage{
\method{predict}{msic}(object, X, ...)
}
\arguments{
\item{object}{An \code{msic} object.}

\item{X}{A numeric matrix of incidence covariates; each row is an observation vector.}

\item{\dots}{Additional arguments.}
}
\value{
A numeric vector of the probability.
}
\description{
Predict the uncure probability given covariate \eqn{X}.
}
