% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msic.R
\name{print.msic}
\alias{print.msic}
\title{Print a msic object}
\usage{
\method{print}{msic}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An \code{msic} object.}

\item{digits}{An integer of significant digits in printout.}

\item{\dots}{Additional print arguments.}
}
\value{
A printout mentioned in Details.
}
\description{
Print a summary of the fitted monotone single-index cure model.
}
\details{
The estimated incidence regression parameter \eqn{\hat{\gamma}} is printed,
followed by the estimated latency regression parameter \eqn{\hat{\beta}}.
}
