% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{smooth.monotone.link}
\alias{smooth.monotone.link}
\title{Smooth monotone link function}
\usage{
smooth.monotone.link(index, mlink, bw)
}
\arguments{
\item{index}{A numeric vector (length \code{n}) of index \eqn{\gamma^TX}.}

\item{mlink}{A \code{\link[stats]{stepfun}} of the monotone link estimate, see \code{\link{em.link}}.}

\item{bw}{A positive number of the bandwidth parameter.}
}
\value{
A numeric vector of the smoothed link values at \code{index}.
}
\description{
Smooth monotone link function, \eqn{\hat{\varphi}_{n,\theta}^s}
}
\keyword{internal}
