% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{smoothing}
\alias{smoothing}
\title{Kernel smoothing of monotone link function using tri-weight kernel}
\usage{
smoothing(x, jump.x, jump.size, bw)
}
\arguments{
\item{x}{A numeric vector of points at which the smoothing is computed.}

\item{jump.x}{A numeric vector of points at which there is a jump.}

\item{jump.size}{A numeric vector of jump heights.}

\item{bw}{A positive number of the bandwidth parameter.}
}
\value{
A numeric vector of the smoothed link values at \code{x}.
}
\description{
Kernel smoothing of monotone link function using tri-weight kernel
}
\keyword{internal}
