% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{uncure.status}
\alias{uncure.status}
\title{Compute the conditional expectation of uncured status}
\usage{
uncure.status(delta, p, s.u, y, y.max = max(y[delta == 1]))
}
\arguments{
\item{delta}{A numeric vector (length \code{n}) of the censoring indicator.}

\item{p}{A numeric vector (length \code{n}) of the uncured probability values.}

\item{s.u}{A numeric vector (length \code{n}) of the
conditional survival function values, see \code{\link{estimate.survival}}.}

\item{y}{A numeric vector (length \code{n}) of the observed event time.}

\item{y.max}{y.max A positive number of the cure threshold.
Default is the maximum uncensored event time.}
}
\value{
A numeric vector of the conditional expected uncured status.
}
\description{
Compute \eqn{\mathbb{E}[B | Y, \Delta, X, Z; \hat{\theta}, \hat{\varphi}]},
where \eqn{\hat{\theta}} and \eqn{\hat{\varphi}} are estimates from the previous iteration.
}
\keyword{internal}
